+++ internal/constants/flags.go
@@ -1,15 +1,24 @@
 package constants
 
 const (
-	HeaderFlagName      = "header"
-	AWSSigV4FlagName    = "aws-sigv4"
-	AWSProfileFlagName  = "aws-profile"
-	AWSRegionFlagName   = "aws-region"
-	JSONBodyFlagName    = "json"
-	OutputFileFlagName  = "output-file"
-	RunTargetFlagName   = "target"
+	HeaderFlagName = "header"
+
+	AWSSigV4FlagName   = "aws-sigv4"
+	AWSProfileFlagName = "aws-profile"
+	AWSRegionFlagName  = "aws-region"
+
+	BodyFlagName             = "body"
+	ResponseBodyOnlyFlagName = "response-body-only"
+	OutputFileFlagName       = "output"
+
+	RunTargetFlagName = "target"
+
 	SandboxFlagName     = "sandbox"
 	SandboxPortFlagName = "port"
-	DetailsFlagName     = "details"
-	TimeoutFlagName     = "timeout"
+
+	DetailsFlagName = "details"
+
+	TimeoutFlagName = "timeout"
+
+	VerboseFlagName = "verbose"
 )