+++ CHANGELOG.md
@@ -4,6 +4,15 @@ This changelog adheres to semantic version according to [keepachangelog](https:/
 
 ## Unreleased
 
+### Added
+- Flag named `--verbose` to output debug logs.
+- Flag named `--response-body-only` to only output the response body.
+
+### Changed
+- Name of `--json` to `--body` in `httpreq post` command. It also changed the behavior to also
+  accept a filename instead of just a string of content.
+- Name of `--output-file` to `--output` only.
+
 ## [0.6.0] - 2019-12-17
 
 ### Added